refactor(offline-sync): digest cache lives in offline-sync-file-io (ratchet)
The bounded plain-file digest helper moves to the io sibling that owns
the plain chunk reader — keeps offline-sync.ts under its LOC ceiling.
